VGA Controller
A VGA controller using a Digilent Basys 3 FPGA.
The implementation of this project was done with Vivado and Verilog HDL code.
The colors are adjustable using 4 bit hexidecimal.
VGAwcolor is the source code and tester is the test bench.
Lab explanantion and process can be found in lab 4 and lab 4 report PDFs.